Move sftp to the apps dir
Also fixes multiple regressions and test cases introduced in series 4
