+++ tardis/apps/sftp/apps.py
@@ -2,5 +2,5 @@ from tardis.app_config import AbstractTardisAppConfig
 
 
 class SFTPConfig(AbstractTardisAppConfig):
-    name = 'tardis.sftp'
+    name = 'tardis.apps.sftp'
     verbose_name = 'SFTP'
