+++ tardis/apps/sftp/templates/sftp/index.html
@@ -51,7 +51,7 @@
             'More Options')<br/>
             Make sure the top dropdown menu reads SFTP<br/>
             <img width="713" height="641"
-                 src="{% url "tardis.sftp:cyberduck_connection_window" %}" />
+                 src="{% url "tardis.apps.sftp:cyberduck_connection_window" %}" />
         </p>
 {#        <h2>SSHFS on Linux</h2>#}
