+++ tardis/default_settings/apps.py
@@ -23,5 +23,6 @@ INSTALLED_APPS = (
     'tardis.analytics',
     # these optional apps, may require extra settings
     'tardis.apps.oaipmh',
+    'tardis.apps.sftp',
     # 'tardis.apps.push_to',
 )
