+++ tardis/urls/__init__.py
@@ -94,9 +94,6 @@ urlpatterns = [
     # Facility views
     url(r'^facility/', include(facility_urls)),
 
-    # SFTP views
-    url(r'^sftp/', include('tardis.sftp.urls', namespace='sftp')),
-
     # Display Views
     url(r'^display/', include(display_urls)),
